requirements.txt
python-telegram-bot==20.8
solana==0.30.1
httpx==0.23.0  # Solana kutubxonasi uchun mos versiya
requests
